preparing build scripts

+++ build_www.py
@@ -5,14 +5,21 @@ import os import chdir
 
 Import("env")
 
+def defined(flag):
+    my_flags = env.ParseFlags(env["BUILD_FLAGS"])
+    for define in my_flags.get("CPPDEFINES"):
+        if (define == flag or (isinstance(define, list) and define[0] == flag)):
+            return True
+            
 def build_web():
     chdir("interface")
     print("Building www...")
     try:
         print(check_output(["npm", "install"]))
         print(check_output(["npm", "run", "build"]))
-        rmtree("../data/www")
-        copytree("build", "../data/www")
+        if not defined("PROGMEM_WWW"):
+            rmtree("../data/www")
+            copytree("build", "../data/www")
     finally:
         chdir("..")
